Made OutputObjectBean that does so we only gets the Data object out and changed in the GetAll call

+++ src/main/java/pocmongodb/util/CVRObject.java
@@ -1,17 +1,17 @@
 package pocmongodb.util;
 
 public class CVRObject {
-	private int _id;
+	private long _id;
 	private long timestamp;
 	
 	public CVRObject() {
 	}
 	
-	public int get_id() {
+	public long get_id() {
 		return _id;
 	}
 	
-	public void set_id(int cvrNumber) {
+	public void set_id(long cvrNumber) {
 		this._id = cvrNumber;
 	}
